+++ django/applications/catmaid/static/libs/catmaid/models/similarity.js
@@ -92,12 +92,13 @@
   };
 
   Similarity.addConfigFromScoring = function(projectId, name, distanceBreaks,
-      dotBreaks, scoring) {
+      dotBreaks, scoring, tangentNeighbors = 20) {
     let params = {
       name: name,
       distance_breaks: distanceBreaks,
       dot_breaks: dotBreaks,
       scoring: scoring,
+      tangent_neighbors: tangentNeighbors,
     };
     return CATMAID.fetch(project.id + '/similarity/configs/', 'PUT', params)
       .then(function(result) {